Install the colorhug command line tool as /usr/bin/colorhug-cmd
Debian and other distros don't have libexec, so put the CLI tool in bindir and
add a proper man page.

Also add a symlink to /usr/libexec/colorhug so the existing documentation is not
wrong. I'll update the webpages to the new path after the next LiveCD release
and the colorhug-client packages have hit distros.